extracted htc_monitor models
